pyproject.toml
[project]
name = "fwii"
version = "0.1.0"
description = "West of England Regional Flood Warning Intensity Index"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "polars>=1.32.0",
    "httpx>=0.27.0",
    "pyyaml",
    "pyarrow",
    "odfpy>=1.4.1",
    "pandas>=3.0.1",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "ruff>=0.9.3",
]

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
fwii-pipeline = "fwii.cli:main_pipeline"
fwii-calculate = "fwii.cli:main_calculate"
fwii-trend = "fwii.cli:main_trend"
